Change to float array, integrate imu (#45)
* For loop works, but only one iteration

* For loops work

* Simple sensor if statements work

* Improved sensor matches

* Change to 9-byte structure

* Updated tests to 9-byte

* 10-bit

* Update cpp-parser.test.ts

* Switch to a float32 array

* Character stack

* Cleanup

* Update bytecode-types.ts

* Update bytecode-types.ts

+++ src/classes/esp32/esp32-socket-manager.ts
@@ -273,8 +273,8 @@ export default class Esp32SocketManager extends Singleton {
 		)
 	}
 
-	public async emitBytecodeToPip(pipUUID: PipUUID, byteCode: Uint8Array): Promise<void> {
-		return await this.emitSocketCommand<Uint8Array>(
+	public async emitBytecodeToPip(pipUUID: PipUUID, byteCode: Float32Array): Promise<void> {
+		return await this.emitSocketCommand<Float32Array>(
 			pipUUID,
 			this.sendEsp32MessageManager.sendBytecodeToPip.bind(this.sendEsp32MessageManager),
 			byteCode,
